Resumeanalyzer/requirements.txt
openai
python-docx
PyMuPDF
Flask==2.0.1
azure-storage-blob==12.9.0
Werkzeug ==2.0.0
mysql-connector-python
